- Wups.
git-svn-id: https://subversion.assembla.com/svn/mapler-me@529 7cbc8f23-9b70-4854-b091-be65f742de0c

+++ branches/MPLRServer/DataParsers/CharacterClasses/CharacterInventory.cs
@@ -392,7 +392,7 @@ namespace MPLRServer
         {
             if (pValue.HasFlag(pFlag))
             {
-                var val = pPacket.ReadShort();
+                var val = pPacket.ReadByte();
                 if (pLogIfFound)
                 {
                     Logger.WriteLine("Found flag {0:X8}: {1}", pFlag, val);
@@ -421,7 +421,7 @@ namespace MPLRServer
         {
             if (pValue.HasFlag(pFlag))
             {
-                var val = pPacket.ReadShort();
+                var val = pPacket.ReadInt();
                 if (pLogIfFound)
                 {
                     Logger.WriteLine("Found flag {0:X8}: {1}", pFlag, val);
